move parse_version
